Refactor component and delete unused files
